src/requirements.txt
anyio==3.6.1
backports.entry-points-selectable==1.1.0
CacheControl==0.12.6
cachy==0.3.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cleo==0.8.1
click==8.1.3
clikit==0.6.2
crashtest==0.3.1
cryptography==3.4.8
distlib==0.3.2
fastapi==0.78.0
filelock==3.0.12
h11==0.13.0
html5lib==1.1
httptools==0.4.0
idna==3.2
jeepney==0.7.1
keyring==21.8.0
lockfile==0.12.2
msgpack==1.0.2
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
pipenv==2021.5.29
pkginfo==1.7.1
platformdirs==2.3.0
poetry==1.1.8
poetry-core==1.0.4
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycparser==2.20
pydantic==1.9.1
pylev==1.4.0
pyparsing==2.4.7
python-dotenv==0.20.0
PyYAML==6.0
qbittorrent-api==2022.5.32
requests==2.26.0
requests-toolbelt==0.9.1
SecretStorage==3.3.1
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tomlkit==0.7.2
typing_extensions==4.3.0
urllib3==1.26.6
uvicorn==0.18.2
uvloop==0.16.0
virtualenv==20.7.2
virtualenv-clone==0.5.6
watchfiles==0.15.0
webencodings==0.5.1
websockets==10.3
